backend/requirements.txt
# YATAV Training System Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.1
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# AI & ML Services
openai>=1.10.0
anthropic>=0.8.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# Audio Processing (STT/TTS)
SpeechRecognition==3.10.1
pydub==0.25.1
whisper==1.1.10
elevenlabs==0.2.26
azure-cognitiveservices-speech==1.34.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
Pillow==10.2.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Production
gunicorn==21.2.0